+++ GANDLF/models/vgg.py
@@ -4,7 +4,6 @@ Modified from https://github.com/pytorch/vision.git
 '''
 import math
 import torch.nn as nn
-
 from .modelBase import get_final_layer
 
 __all__ = [
@@ -17,9 +16,10 @@ class VGG(nn.Module):
     '''
     VGG model 
     '''
-    def __init__(self, n_dimensions, features, n_outputClasses, final_convolution_layer: str = 'softmax',):
+    def __init__(self, n_dimensions, features, inputFeaturesForClassifier, n_outputClasses, final_convolution_layer: str = 'softmax',):
         super(VGG, self).__init__()
         self.features = features
+        self.final_convolution_layer = get_final_layer(final_convolution_layer)
         if n_dimensions == 2:
             self.Conv = nn.Conv2d
         elif n_dimensions == 3:
@@ -27,15 +27,14 @@ class VGG(nn.Module):
 
         self.classifier = nn.Sequential(
             nn.Dropout(),
-            nn.Linear(512, 512),
+            nn.Linear(inputFeaturesForClassifier, 512),
             nn.ReLU(True),
             nn.Dropout(),
             nn.Linear(512, 512),
             nn.ReLU(True),
             nn.Linear(512, 10),
             nn.ReLU(True),
-            nn.Linear(10, n_outputClasses),
-            get_final_layer(final_convolution_layer)
+            nn.Linear(10, n_outputClasses) 
         )
          # Initialize weights
         for m in self.modules():